fix : notification can open up in all path (#29)
